requirements.txt
botbuilder-core>=4.7.0
aiohttp
cryptography==38.0.1